👌 IMPROVE: Rename jsonobject module to json_elements
